+++ skoda-tender-ios/Domain/Model/Item.swift
